// src/enums/strikePriceInterpretations.enum.ts
// eslint-disable-next-line no-shadow
export enum StrikePriceInterpretationsEnum {
  SELLING_CONTRACTS = 'sellingContracts',
  SQUARE_OFF_SOLD_CONTRACTS = 'squareOffSoldContracts',
  BUYING_CONTRACTS = 'buyingContracts',
  SQUARE_OFF_BOUGHT_CONTRACTS = 'squareOffBoughtContracts',
  NO_SIGNIFICANT_MOVEMENT = 'noSignificantMovement',
}